requirements.txt
backpack==0.1
blinker==1.5
cleo==0.6.8
Faker==0.8.18
gunicorn==20.1.0
inflection==0.3.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
pyaml==16.12.2
Pygments==2.13.0
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pytzdata==2020.1
PyYAML==5.4.1
simplejson==3.17.6
six==1.16.0
text-unidecode==1.2
tzlocal==1.5.1
wrapt==1.14.1
